Use the system provided harfbuzz on chromeos
It's part of the pango library we use already, so by using it, we avoid
a link time collission.

However, if you build just the browser on linux on ubuntu 12, we don't
have such an up to date pango available, so we need to be a bit smart
about whether or not to use the system harfbuzz.

This also allows for treating linker warnings as errors on chromeos

BUG=353127
R=jshin@chromium.org,thakis@chromium.org

Review URL: https://codereview.chromium.org/203163003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a281d79e4d50bf97e0b9d03166a8c46963a30864