Add unpatched OpenBSD trees to tools/android/heap_profiler/
OpenBSD trees are the most convenient C implementation
of RB-Trees (just one header file).
This is needed for libheap_profiler, a library for profiling
memory allocations on Android (and probably later on Linux)
which must be a dependency-free.
Note: the target library is not mean to be shipped as part
of Chrome. It is a separate tool which can be build
and deployed separately for debugging/profiling purposes.

BUG=382489
NOTRY=true

Review URL: https://codereview.chromium.org/323873004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c418c17769a00b8590f524563c0a66056b72ff3d