Compile braille ime among other compiled js resources.
This includes the braille IME in the compiled js resources.  As a cleanup,
most of the braille IME externs are removed, consolidating with the
central location in third_party/closure_compiler.  The compiler
used for checking braille IME and chroemvox in the presubmit scripts is
chagned to the one in top-level third_party, isntead of the one
in blink.

R=dtseng, tbreisacher
BUG=

Review URL: https://codereview.chromium.org/665633002

Cr-Original-Commit-Position: refs/heads/master@{#301327}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d6c36e32a81f912a3722238444c27fc6dd5d7bb8