[sql] Import reference version of SQLite 3.7.6.3.
In working on importing a newer version of SQLite, I found that I could
not find a copy of the previous version.  This makes it harder to figure
out if we have any mods which need to be tracked or unwound.

This import should be a non-event, since most of the files are identical
to the files in third_party/sqlite/src/ , so git will just add a reference.

BUG=340757

Review URL: https://codereview.chromium.org/826543003

Cr-Original-Commit-Position: refs/heads/master@{#311378}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9cf3c6f4c0a79991a5ebacebee62adb0fb267861