IndexedDB: Revert change to enable reuse of leveldb manifest in IDB (take 2).
The prior revert (ca6c875) of the leveldb manifest reuse change (de701ab) was
insufficient because there appears to be a bug in leveldb where it fails to
consult the reuse_logs option. So if the Env has a working NewAppendableFile
method then leveldb will always try to reuse it's manifest. Deleting our
version of this function so that the base Env implementation will be called
(which fails forcing no-reuse).

BUG=402980, 395799

Review URL: https://codereview.chromium.org/852073002

Cr-Original-Commit-Position: refs/heads/master@{#311576}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 617e94e4e3b423579ecebc1c01e07bfbd603694d