IndexedDB: Enable leveldb log reuse on Chrome OS.
Reuse was disabled in #317499 because CrOS corruption rates were spiking and
were too close to the stable branch point to gather more history.  Re-enabling
to allow monitoring in dev/beta channels with the hope that the rise we saw on
CrOS was normal fluctuations and unrelated to this change.

BUG=460568

Review URL: https://codereview.chromium.org/1023773003

Cr-Original-Commit-Position: refs/heads/master@{#321427}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 319835313295e85743ad4f3e4f45af4595a53033