mac: Prepare apple_sample_code for compiling against the OSX 10.10 SDK.
The code emitted a warning when compiled against the OSX 10.10 SDK due to the
introduction of a new enum in OSX 10.10. This CL defines the enum when
compiling against an older SDK and uses the new enum in place of NSUInteger.

BUG=463170

Review URL: https://codereview.chromium.org/1121513002

Cr-Original-Commit-Position: refs/heads/master@{#327767}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8ad8099174b10bb15889909a67af27d90096817f