qcms: Limit vcgt table to a maximum of 1024 entries
This is a reasonable limit on the size of the vcgt to permit; this will
prevent rogue data potentially causing an OOM situation.

TEST=With test file blue-vcgt-test-color-profile-1024.icc observe that
the vcgt data is parsed and made available through the API but with
blue-vcgt-test-color-profile-2048.icc the vcgt is ignored.
R=noel@chromium.org
BUG=471749

Review URL: https://codereview.chromium.org/1138473002

Cr-Original-Commit-Position: refs/heads/master@{#328966}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4cc1320a9ba15971f5d67f704b5e732d61ba2cdd